products/video-ad-generator/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
sqlalchemy>=2.0.0
anthropic>=0.40.0
python-telegram-bot>=21.0
google-genai>=1.52.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0
